requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
atlassian-python-api==3.41.2
msal==1.26.0
msgraph-core==1.0.0 